Improved coverage for coordinator, up to 93%, split that test file into smaller files
