+++ tests/unit/core/coordinator/__init__.py
@@ -1,0 +1,3 @@
+"""
+Coordinator test package initialization.
+"""
